Still working on issue #70. - Solved problem with image sizing on databases page on IE8 - Questionaries visualization and editing areas are now accessible on IE8. They previously weren’t. They are not fully functional yet, though.
The problem was the way IE8 interprets relative addresses and the way the addressment was made in pages deeper in the tree.
It needs the “/“ at the start to consider the addresses relative to the root (and since IE8 doesn’t support the push states, for him, addresses are always complete addresses, not ways to pass data, and must be fully specified “correctly”). It now works