- Removed bootstrap_ie_compatibility, from static files, because now its generated dynamically because it includes paths inside - Small corrections here an there. Errors that most browsers don’t complain about except for IE (missing apostrophes, etc) - Removed duplicate query inclusion (some pages need it to be on top, so there’s no need to 2 inclusions, it can always be on top). - Small correction on questionset.js. On ie8 e 7 the method getAttribute() doesn’t return correctly, replaced by jquery equivalent $().attr()
On IE., we already have, more or less CRUD operations.

Small problem remains though, like referes on issue #70. Problems with gigantic questionaries being slow. This probably due to bad performance of the IE 8 javascript engine. Tried it on a virtual machine with 1 core and 512 mb and it was not very usable. Will try it with a real setup to check better.

Maybe a good path will be trying to improve performance, or reduce the number of questions on memory at a time, and do it by parts.