- While analysing the code for tablediff widget, i found it was not hiding empty question sets (removed through filtering by removal of unmatched/matched/empty, etc). This can be quite a bother for a user looking for similarities that had to look to a lot of empty container, i added a function that verifies the state of that and hides them when they are empty. - Removed the limiter on the results table. - Everything seems to be fine server side to support this feature i didn't need to change anything.
To Do:
- I will start looking into the setting of the base reference database so all are compared to the one, and the one can be changed.
- I will start looking into the filtering by search functionality (this seems easier i thought it would be).

At first glance, everything is done using javascript, there's a bit of a performance problem, that indicates i will probably have to have this comparisons be made other way to better scalate to a lot of databases. Maybe instead of dom manipulation, load everything to a javascript list on page load, compare it on javascript, and only then load it to the page (this could be done by blocks, and using the thread pool approach we are using in other places). And would probably be faster, since one of the slowest operations in javascript are dom manipulations.