Still experimenting on #41:
Now the partial saving on changing question sets is incremental and a lot faster., so this seems to be the way.

The problem remains the same though, the mandatory parameters being all over the place while we only save a little part. I didn't found a way to get around those, yet.