Fixed #387:
- Added html templates to emails from django userena

ATTENTION: This involved updating django-userena to the most recent version, since the old version we were using didn't support it. Since we have a forked version i had to merge the master from the django-userena base against ours, and resolve the conflicts.

So this update will only work after updating the fork that is a dependency on this project.