Fixed compression o #214: - Added django-compressor. - Fixed pages so django compressor works (had to create a separate block for head css includes, named styleinclude, and move css file includes on headinclude to there). Also had to fix so js files with error that broke django-compressor (like missing ;)
the appliance of compression is based by default on debug variable, but can be force throughout the usage of "COMPRESS_ENABLED" on settings.py. For demonstration proposes on the pull request i activated it, but we probably should work with it deactivated while on development and use the default settings.