Fixed #610: - Fully working import/export solution for fingerprint information, everything but private shares is being replicated.
Also:
- Fixed bug with redirect middleware on debug mode (this doesn't affect
production, but was bugging me, it makes sense if we ever wanted to run
the project from a flat directory anyway)
- Fixed bug with literature not showing since refactor

How to use:

To export do:

    python manage.py fingerprint_export <hash> directory of destiny

To import do:

    python manage.py fingerprint_import <path_to_.emif_file>    


The commands rely on the existence of the questionnaire, it will fail if
the model doesn't exist. It also relies on the existence of the main
owner.

Any other user (like in the comments) that doesn't exist on the system
the file is imported to, simply stays as anonymous.

The command of import will try to maintain consistency with the
identification numbers and hashes, however, if it detects the hash
already exists, it procedes to add the database with a new hash and ids
for the components, renumbering everything. This acts kind of like a
duplicate function in this case.