Fixed #418: - Added the capability of defining restricted users.
This users can only see a list of hand-defined databases(and also their
own database creations).

The user can be defined as a restricted user on django-admin > Accounts
> EmifProfile

After a user is restricted, the databases can be added through
django-admin > Accounts RestrictedUserDbs table.

Each database the restricted user creates, is added to the list of
allowed databases.
Each database that is shared with the restricted user, is added to the
list of allowed databases.


To deploy, there's a need to migrate accounts app, please do:

    python manage.py migrate accounts