Fixing #831: - Changed static emif catalogue reference to dynamic reference.
This is a bit bothersome since requires the change of all email_template
models. Probably the fastest way on deploys already running is going to
the templates and doing CTRL+F "Emif Catalogue" and replacing all
references with {{site.name}}, which is basically what is done in the
sql script.